refactoring: OCP 원칙 고려 (Redis)
