+++ src/main/java/com/unionclass/memberservice/common/redis/redisLockUtils/RedisLockUtils.java
@@ -1,0 +1,4 @@
+package com.unionclass.memberservice.common.redis.redisLockUtils;
+
+public interface RedisLockUtils {
+}
