+++ src/main/java/com/unionclass/memberservice/common/redis/redisLockUtils/RedisLockUtilsImpl.java
@@ -1,0 +1,11 @@
+package com.unionclass.memberservice.common.redis.redisLockUtils;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+public class RedisLockUtilsImpl implements  RedisLockUtils {
+
+
+}
